[native_toolchain_c] Add linking for Android
