+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_android_test.dart
@@ -7,7 +7,6 @@ import 'dart:io';
 import 'package:code_assets/code_assets.dart';
 import 'package:hooks/hooks.dart';
 import 'package:native_toolchain_c/native_toolchain_c.dart';
-import 'package:native_toolchain_c/src/utils/run_process.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';
@@ -20,14 +19,6 @@ void main() {
     Architecture.x64,
     Architecture.riscv64,
   ];
-
-  const objdumpFileFormat = {
-    Architecture.arm: 'elf32-littlearm',
-    Architecture.arm64: 'elf64-littleaarch64',
-    Architecture.ia32: 'elf32-i386',
-    Architecture.x64: 'elf64-x86-64',
-    Architecture.riscv64: 'elf64-littleriscv',
-  };
 
   /// From https://docs.flutter.dev/reference/supported-platforms.
   const flutterAndroidNdkVersionLowestSupported = 21;
@@ -58,21 +49,7 @@ void main() {
             linkMode,
             optimizationLevel: optimizationLevel,
           );
-          if (Platform.isLinux) {
-            final machine = await readelfMachine(libUri.path);
-            expect(machine, contains(readElfMachine[target]));
-          } else if (Platform.isMacOS) {
-            final result = await runProcess(
-              executable: Uri.file('objdump'),
-              arguments: ['-T', libUri.path],
-              logger: logger,
-            );
-            expect(result.exitCode, 0);
-            final machine = result.stdout
-                .split('\n')
-                .firstWhere((e) => e.contains('file format'));
-            expect(machine, contains(objdumpFileFormat[target]));
-          }
+          await expectMachineArchitecture(libUri, target);
           if (linkMode == DynamicLoadingBundled()) {
             await expectPageSize(libUri, 16 * 1024);
           }
