+++ pkgs/native_toolchain_c/test/clinker/objects_cross_android_test.dart
@@ -2,31 +2,20 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-//TODO(mosuem): Enable for windows and mac.
-// See https://github.com/dart-lang/native/issues/1376.
-@TestOn('linux')
-library;
-
-import 'dart:io';
-
 import 'package:code_assets/code_assets.dart';
-import 'package:test/test.dart';
 
-import 'treeshake_helper.dart';
+import 'objects_helper.dart';
 
 Future<void> main() async {
-  if (!Platform.isLinux) {
-    // Avoid needing status files on Dart SDK CI.
-    return;
-  }
-
   final architectures = [
     Architecture.arm,
     Architecture.arm64,
     Architecture.ia32,
     Architecture.x64,
     Architecture.riscv64,
-  ]..remove(Architecture.current);
+  ];
 
-  await runTreeshakeTests(OS.current, architectures);
+  const targetOS = OS.android;
+
+  await runObjectTests(targetOS, architectures);
 }
