+++ pkgs/native_toolchain_c/test/clinker/treeshake_helper.dart
@@ -2,11 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-//TODO(mosuem): Enable for windows and mac.
-// See https://github.com/dart-lang/native/issues/1376.
-@TestOn('linux')
-library;
-
 import 'dart:io';
 
 import 'package:code_assets/code_assets.dart';
@@ -17,13 +12,16 @@ import 'package:test/test.dart';
 import '../helpers.dart';
 import 'build_testfiles.dart';
 
-Future<void> runTests(List<Architecture> architectures) async {
+Future<void> runTreeshakeTests(
+  OS targetOS,
+  List<Architecture> architectures,
+) async {
   CLinker linkerManual(List<String> sources) => CLinker.library(
     name: 'mylibname',
     assetName: '',
     sources: sources,
     linkerOptions: LinkerOptions.manual(
-      flags: ['--strip-debug', '-u', 'my_other_func'],
+      flags: ['--strip-debug', '-u=my_other_func'],
       gcSections: true,
       linkerScript: packageUri.resolve(
         'test/clinker/testfiles/linker/symbols.lds',
@@ -43,8 +41,6 @@ Future<void> runTests(List<Architecture> architectures) async {
     linkerOptions: LinkerOptions.treeshake(symbols: null),
   );
 
-  const os = OS.linux;
-
   late Map<String, int> sizes;
   sizes = <String, int>{};
   for (final architecture in architectures) {
@@ -53,65 +49,63 @@ Future<void> runTests(List<Architecture> architectures) async {
       (name: 'auto', linker: linkerAuto),
       (name: 'autoEmpty', linker: linkerAutoEmpty),
     ]) {
-      test(
-        'link test with CLinker ${clinker.name} and target $architecture',
-        () async {
-          final tempUri = await tempDirForTest();
-          final tempUri2 = await tempDirForTest();
-          final testArchive = await buildTestArchive(
-            tempUri,
-            tempUri2,
-            os,
-            architecture,
+      test('link test with CLinker ${clinker.name} and target '
+          '$architecture for targetOS $targetOS', () async {
+        final tempUri = await tempDirForTest();
+        final tempUri2 = await tempDirForTest();
+        final testArchive = await buildTestArchive(
+          tempUri,
+          tempUri2,
+          targetOS,
+          architecture,
+        );
+
+        final linkInputBuilder = LinkInputBuilder()
+          ..setupShared(
+            packageName: 'testpackage',
+            packageRoot: tempUri,
+            outputFile: tempUri.resolve('output.json'),
+            outputDirectoryShared: tempUri2,
+          )
+          ..setupLink(assets: [], recordedUsesFile: null)
+          ..addExtension(
+            CodeAssetExtension(
+              targetOS: targetOS,
+              targetArchitecture: architecture,
+              linkModePreference: LinkModePreference.dynamic,
+              cCompiler: cCompiler,
+              android: targetOS == OS.android
+                  ? AndroidCodeConfig(targetNdkApi: 21)
+                  : null,
+            ),
           );
 
-          final linkInputBuilder = LinkInputBuilder()
-            ..setupShared(
-              packageName: 'testpackage',
-              packageRoot: tempUri,
-              outputFile: tempUri.resolve('output.json'),
-              outputDirectoryShared: tempUri2,
-            )
-            ..setupLink(assets: [], recordedUsesFile: null)
-            ..addExtension(
-              CodeAssetExtension(
-                targetOS: os,
-                targetArchitecture: architecture,
-                linkModePreference: LinkModePreference.dynamic,
-                cCompiler: cCompiler,
-              ),
-            );
+        final linkInput = linkInputBuilder.build();
+        final linkOutputBuilder = LinkOutputBuilder();
 
-          final linkInput = linkInputBuilder.build();
-          final linkOutputBuilder = LinkOutputBuilder();
+        printOnFailure(linkInput.config.code.cCompiler.toString());
+        printOnFailure(Platform.environment.keys.toList().toString());
+        await clinker
+            .linker([testArchive.toFilePath()])
+            .run(input: linkInput, output: linkOutputBuilder, logger: logger);
 
-          printOnFailure(linkInput.config.code.cCompiler.toString());
-          printOnFailure(Platform.environment.keys.toList().toString());
-          await clinker
-              .linker([testArchive.toFilePath()])
-              .run(input: linkInput, output: linkOutputBuilder, logger: logger);
+        final linkOutput = linkOutputBuilder.build();
+        final asset = linkOutput.assets.code.first;
 
-          final linkOutput = linkOutputBuilder.build();
-          final asset = linkOutput.assets.code.first;
-          final filePath = asset.file!.toFilePath();
+        await expectMachineArchitecture(asset.file!, architecture);
 
-          final machine = await readelfMachine(filePath);
-          expect(machine, contains(readElfMachine[architecture]));
+        final symbols = await nmReadSymbols(asset);
+        if (clinker.linker != linkerAutoEmpty) {
+          expect(symbols, contains('my_other_func'));
+          expect(symbols, isNot(contains('my_func')));
+        } else {
+          expect(symbols, contains('my_other_func'));
+          expect(symbols, contains('my_func'));
+        }
 
-          final symbols = await nmReadSymbols(asset);
-          if (clinker.linker != linkerAutoEmpty) {
-            expect(symbols, contains('my_other_func'));
-            expect(symbols, isNot(contains('my_func')));
-          } else {
-            expect(symbols, contains('my_other_func'));
-            expect(symbols, contains('my_func'));
-          }
-
-          final du = Process.runSync('du', ['-sb', filePath]).stdout as String;
-          final sizeInBytes = int.parse(du.split('\t')[0]);
-          sizes[clinker.name] = sizeInBytes;
-        },
-      );
+        final sizeInBytes = await File.fromUri(asset.file!).length();
+        sizes[clinker.name] = sizeInBytes;
+      });
     }
     tearDownAll(() {
       expect(
