+++ pkgs/native_toolchain_c/test/helpers.dart
@@ -261,7 +261,7 @@ Future<void> expectSymbols({
   required CodeAsset asset,
   required List<String> symbols,
 }) async {
-  if (Platform.isLinux) {
+  if (Platform.isLinux || Platform.isMacOS) {
     final nmOutput = await nmReadSymbols(asset);
 
     expect(nmOutput, stringContainsInOrder(symbols));
@@ -315,3 +315,29 @@ Future<void> expectPageSize(Uri dylib, int pageSize) async {
 }
 
 int defaultMacOSVersion = 13;
+
+const objdumpFileFormat = {
+  Architecture.arm: 'elf32-littlearm',
+  Architecture.arm64: 'elf64-littleaarch64',
+  Architecture.ia32: 'elf32-i386',
+  Architecture.x64: 'elf64-x86-64',
+  Architecture.riscv64: 'elf64-littleriscv',
+};
+
+Future<void> expectMachineArchitecture(Uri libUri, Architecture target) async {
+  if (Platform.isLinux) {
+    final machine = await readelfMachine(libUri.path);
+    expect(machine, contains(readElfMachine[target]));
+  } else if (Platform.isMacOS) {
+    final result = await runProcess(
+      executable: Uri.file('objdump'),
+      arguments: ['-T', libUri.path],
+      logger: logger,
+    );
+    expect(result.exitCode, 0);
+    final machine = result.stdout
+        .split('\n')
+        .firstWhere((e) => e.contains('file format'));
+    expect(machine, contains(objdumpFileFormat[target]));
+  }
+}
